moved workflowOptions to its own file in workflowOptions.ts
