+++ src/workflow.ts
@@ -3,7 +3,8 @@ import prompts from "prompts";
 import { Selections } from "./types.js";
 import { getUserPackageManager } from "./utils/helpers.js";
 import { validateProjectName } from "./utils/validation.js";
-import { rechoseWorkflow, workflowOptions } from "./rechoseWorkflow.js";
+import { rechoseWorkflow } from "./rechoseWorkflow.js";
+import { workflowOptions } from "./workflowOptions.js";
 
 export async function startWorkflow() {
   let initialResult: prompts.Answers<
